+++ README.md
@@ -3,3 +3,12 @@
 В директории `tests` находятся тесты для проверки API, которое должно быть реализовано в веб-сервере.
 
 Директория `web` содержит файлы фронтенда.
+go test -run ^TestDB$ ./tests
+go test -run ^TestNextDate$ ./tests
+go test -run ^TestAddTask$ ./tests
+go test -run ^TestTasks$ ./tests
+go test -run ^TestEditTask$ ./tests
+go test -run ^TestDone$ ./tests
+go test -run ^TestDelTask$ ./tests
+go test ./tests
+
